# openapi.yaml
openapi: 3.1.0
info:
  title: Story Readiness API
  summary: Check whether a user story has enough information to start implementation.
  description: 'Each story is judged by TypeSafe''s Jev model in one request. Code then applies blocking
    rules, weights, and thresholds to reach a verdict. Checks depend on the story type: a persona is required
    for user features only, a statement of value is not required for bugs, and technical stories are judged
    on safe rollout instead of application edge cases.


    Weights and thresholds are unvalidated starting points; see the README.'
  version: 0.1.0
paths:
...
  /api/boards/{board_id}/stories:
    get:
      tags:
      - stories
      summary: List a board's stored stories with their latest assessment
      operationId: listStories
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StoredStoryOut'
                type: array
                title: Response Liststories
        '404':
          description: The board was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - stories
      summary: Store a pasted story on a board
      description: Adds the story to the end of the board's backlog. A board holds at most 100 stories, counting
        every column.
      operationId: createStory
      parameters:
      - name: board_id
        in: path
        required: true
        schema:
          type: string
          title: Board Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryIn'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredStoryOut'
        '404':
          description: The board was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: The board already holds the most stories it can.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '413':
          description: The story is longer than the size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stories/{story_id}:
    get:
      tags:
      - stories
      summary: Get a stored story with its latest report and assessment history
      operationId: getStory
      parameters:
      - name: story_id
        in: path
        required: true
        schema:
          type: string
          title: Story Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredStoryDetailOut'
        '404':
          description: The story was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - stories
      summary: Replace a stored story's text
      description: Earlier assessments are kept; the story is marked stale until it is assessed again.
      operationId: updateStory
      parameters:
      - name: story_id
        in: path
        required: true
        schema:
          type: string
          title: Story Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredStoryOut'
        '404':
          description: The story was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '413':
          description: The story is longer than the size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - stories
      summary: Delete a stored story and its assessment history
      operationId: deleteStory
      parameters:
      - name: story_id
        in: path
        required: true
        schema:
          type: string
          title: Story Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: The story was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stories/{story_id}/move:
    put:
      tags:
      - stories
      summary: Move a stored story to a workflow column and position
      description: Used by drag and drop on the board. Does not change the story's text or mark it stale. Moving to the blocked column requires a blocked_reason; moving anywhere else clears it.
      operationId: moveStory
      parameters:
      - name: story_id
        in: path
        required: true
        schema:
          type: string
          title: Story Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoryMoveIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredStoryOut'
        '404':
          description: The story was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/stories/{story_id}/assess:
    post:
      tags:
      - stories
      summary: Assess a stored story and save the report to its history
      operationId: assessStoredStory
      parameters:
      - name: story_id
        in: path
        required: true
        schema:
          type: string
          title: Story Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
        '404':
          description: The story was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '413':
          description: The story is longer than the size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '502':
          description: The TypeSafe request failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/webhooks/linear:
    post:
      tags:
      - webhooks
      summary: Receive a Linear Issue webhook
      operationId: linearWebhook
      description: Verifies the HMAC-SHA256 signature and the replay window, then schedules
        processing of the issue. Returns 200 immediately so Linear does not retry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Linear Issue webhook payload.
      responses:
        '200':
          description: Accepted (processing happens asynchronously).
          content:
            application/json:
              schema: {}
        '400':
          description: The signed payload is not a valid webhook body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '401':
          description: Invalid or missing signature, or stale webhook timestamp.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
        '409':
          description: The Linear integration is disabled, or LINEAR_WEBHOOK_SECRET is not configured on the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
components:
  schemas:
...
    AssessmentSummaryOut:
      properties:
        verdict:
          type: string
          enum:
          - ready
          - discuss
          - needs_refinement
          - not_ready
          title: Verdict
        quality:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Quality
        question_count:
          type: integer
          title: Question Count
          description: Questions for the author in this assessment.
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - verdict
      - quality
      - question_count
      - created_at
      title: AssessmentSummaryOut
...
    ErrorOut:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ReportOut:
      properties:
        verdict:
          type: string
          enum:
          - ready
          - discuss
          - needs_refinement
          - not_ready
          title: Verdict
          description: 'First match wins: `not_ready` (a blocker fails), `needs_refinement` (quality below
            0.6), `discuss` (some answers are unsure), `ready`.'
        quality:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Quality
          description: Weighted average of the `weighted` checks.
        checks:
          items:
            $ref: '#/components/schemas/CheckOut'
          type: array
          title: Checks
        story:
          $ref: '#/components/schemas/StoryOut'
        story_type:
          $ref: '#/components/schemas/StoryType'
        blockers_failed:
          items:
            type: string
          type: array
          title: Blockers Failed
          description: Ids of blocker checks that failed confidently.
        to_discuss:
          items:
            type: string
          type: array
          title: To Discuss
          description: Ids of checks whose answers were unsure.
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: TypeSafe model that answered, e.g. `jev-1.13.0`.
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
          description: TypeSafe request id, for support.
        latency_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Latency Ms
          description: Time spent waiting for TypeSafe.
        input_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Input Tokens
      type: object
      required:
      - verdict
      - quality
      - checks
      - story
      - story_type
      - blockers_failed
      - to_discuss
      - model
      - request_id
      - latency_ms
      - input_tokens
      title: ReportOut
...
    StoredStoryDetailOut:
      properties:
        id:
          type: string
          title: Id
        board_id:
          type: string
          title: Board Id
        key:
          type: string
          title: Key
          description: 'Short, human-readable key: the board''s prefix and a number, e.g. FLW-12. Never reused
            after a delete.'
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        acceptance_criteria:
          type: string
          title: Acceptance Criteria
        definition_of_ready:
          items:
            type: string
          type: array
          title: Definition Of Ready
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        status:
          type: string
          enum:
          - backlog
          - refinement
          - ready_for_sprint
          - in_sprint
          - done
          - blocked
          title: Status
          description: Workflow column on the board.
        position:
          type: number
          title: Position
        blocked_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Blocked Reason
          description: Why the story is blocked; set only in the blocked column.
        latest:
          anyOf:
          - $ref: '#/components/schemas/AssessmentSummaryOut'
          - type: 'null'
          description: The latest assessment, or null if never assessed.
        previous_quality:
          anyOf:
          - type: number
          - type: 'null'
          title: Previous Quality
          description: Quality of the assessment before the latest one.
        stale:
          type: boolean
          title: Stale
          description: The story was edited after its latest assessment.
        report:
          anyOf:
          - $ref: '#/components/schemas/ReportOut'
          - type: 'null'
          description: Full report of the latest assessment.
        history:
          items:
            $ref: '#/components/schemas/AssessmentSummaryOut'
          type: array
          title: History
          description: All assessments, newest first.
      type: object
      required:
      - id
      - board_id
      - key
      - title
      - description
      - acceptance_criteria
      - definition_of_ready
      - created_at
      - updated_at
      - status
      - position
      - blocked_reason
      - latest
      - previous_quality
      - stale
      - report
      - history
      title: StoredStoryDetailOut
    StoredStoryOut:
      properties:
        id:
          type: string
          title: Id
        board_id:
          type: string
          title: Board Id
        key:
          type: string
          title: Key
          description: 'Short, human-readable key: the board''s prefix and a number, e.g. FLW-12. Never reused
            after a delete.'
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        acceptance_criteria:
          type: string
          title: Acceptance Criteria
        definition_of_ready:
          items:
            type: string
          type: array
          title: Definition Of Ready
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        status:
          type: string
          enum:
          - backlog
          - refinement
          - ready_for_sprint
          - in_sprint
          - done
          - blocked
          title: Status
          description: Workflow column on the board.
        position:
          type: number
          title: Position
        blocked_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Blocked Reason
          description: Why the story is blocked; set only in the blocked column.
        latest:
          anyOf:
          - $ref: '#/components/schemas/AssessmentSummaryOut'
          - type: 'null'
          description: The latest assessment, or null if never assessed.
        previous_quality:
          anyOf:
          - type: number
          - type: 'null'
          title: Previous Quality
          description: Quality of the assessment before the latest one.
        stale:
          type: boolean
          title: Stale
          description: The story was edited after its latest assessment.
      type: object
      required:
      - id
      - board_id
      - key
      - title
      - description
      - acceptance_criteria
      - definition_of_ready
      - created_at
      - updated_at
      - status
      - position
      - blocked_reason
      - latest
      - previous_quality
      - stale
      title: StoredStoryOut
    StoryIn:
      properties:
        title:
          type: string
          maxLength: 500
          minLength: 1
          title: Title
          description: The story's title or summary.
        description:
          type: string
          maxLength: 10000
          title: Description
          default: ''
          description: The story body. If `acceptance_criteria` is empty, a section under an 'Acceptance criteria'
            heading here is moved into it.
        acceptance_criteria:
          type: string
          maxLength: 10000
          title: Acceptance Criteria
          default: ''
          description: One criterion per bullet, or Given/When/Then scenarios.
        definition_of_ready:
          items:
            type: string
          type: array
          maxItems: 50
          title: Definition Of Ready
          description: Extra requirements from the team's Definition of Ready, one per item. Each becomes
            a yes/no check. Blank and duplicate items are dropped; at most 12 are used, each cut to 300 characters.
      type: object
      required:
      - title
      title: StoryIn
      description: A story to store. Same fields and limits as AssessRequest.
    StoryMoveIn:
      properties:
        status:
          type: string
          enum:
          - backlog
          - refinement
          - ready_for_sprint
          - in_sprint
          - done
          - blocked
          title: Status
          description: Workflow column on the board.
        position:
          type: number
          title: Position
          description: Order within the column, ascending. The client picks a value between the neighbours.
        blocked_reason:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Blocked Reason
          description: Why the story is blocked. Required when status is blocked; not allowed otherwise.
      type: object
      required:
      - status
      - position
      title: StoryMoveIn
    StoryOut:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        acceptance_criteria:
          type: string
          title: Acceptance Criteria
        key:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
          description: Issue key when imported from a tracker.
        source:
          type: string
          enum:
          - paste
          - jira
          - linear
          title: Source
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
          description: Link to the issue when imported from a tracker.
      type: object
      required:
      - title
      - description
      - acceptance_criteria
      - key
      - source
      - url
      title: StoryOut
      description: The story as judged, after acceptance criteria were split out of the description.
    StoryType:
      properties:
        choice:
          type: string
          enum:
          - user_feature
          - technical
          - bug
          title: Choice
          description: The most likely story type.
        confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Confidence
        probabilities:
          additionalProperties:
            type: number
          type: object
          title: Probabilities
          description: Probability of each story type.
      type: object
      required:
      - choice
      - confidence
      - probabilities
      title: StoryType
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...